# Improved ArrayList performance by removing `size` field.