--- a/lists/arraylist/serialization.go
+++ b/lists/arraylist/serialization.go
@@ -16,15 +16,12 @@ var _ containers.JSONDeserializer = (*List[int])(nil)
 
 // ToJSON outputs the JSON representation of list's elements.
 func (list *List[T]) ToJSON() ([]byte, error) {
-	return json.Marshal(list.elements[:list.size])
+	return json.Marshal(list.elements)
 }
 
 // FromJSON populates list's elements from the input JSON representation.
 func (list *List[T]) FromJSON(data []byte) error {
 	err := json.Unmarshal(data, &list.elements)
-	if err == nil {
-		list.size = len(list.elements)
-	}
 	return err
 }