@emirpasic 2024-06-19
@ksw2000 interesting, but expected benchmarks, i.e. the native implementation will always be faster. 

@PapaCharlie what do you think? Tests are passing, if OK, let's merge...