@PapaCharlie 2024-06-19
Gave this a quick once over, I think we can do even better! Specifically around Add. If we just used append, I bet we'd see some nice speedups. Let me open some comments