@ksw2000 2024-06-19
In fact, I tried `append` before making the pull request. The performance of `append()` is worse than `for` loop.

```diff
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
+	list.elements = append(list.elements[:l], values...)
-	for i := range values {
-		list.elements[l+i] = values[i]
-	}
}
```

```
goos: linux
goarch: amd64
pkg: github.com/emirpasic/gods/v2/lists/arraylist
cpu: Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz
                     │ for-loop.txt │              append.txt              │
                     │    sec/op    │    sec/op     vs base                │
ArrayListAdd100-8       788.6n ± 1%    837.4n ± 1%   +6.19% (p=0.000 n=30)
ArrayListAdd1000-8      7.732µ ± 2%   12.390µ ± 6%  +60.25% (p=0.000 n=30)
ArrayListAdd10000-8     76.45µ ± 0%   124.02µ ± 1%  +62.23% (p=0.000 n=30)
ArrayListAdd100000-8    764.5µ ± 2%   1205.2µ ± 3%  +57.64% (p=0.000 n=30)
geomean                 24.43µ         35.29µ       +44.43%

                     │ for-loop.txt │              append.txt              │
                     │     B/op     │     B/op      vs base                │
ArrayListAdd100-8      2.075Ki ± 1%   1.647Ki ± 2%  -20.64% (p=0.000 n=30)
ArrayListAdd1000-8     24.15Ki ± 3%   26.86Ki ± 5%  +11.25% (p=0.009 n=30)
ArrayListAdd10000-8    204.5Ki ± 1%   286.3Ki ± 1%  +40.04% (p=0.000 n=30)
ArrayListAdd100000-8   2.307Mi ± 4%   2.658Mi ± 5%  +15.26% (p=0.000 n=30)
geomean                70.14Ki        76.63Ki        +9.26%

                     │ for-loop.txt │             append.txt              │
                     │  allocs/op   │ allocs/op   vs base                 │
ArrayListAdd100-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd1000-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd10000-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd100000-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
geomean                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```