@PapaCharlie 2024-06-19
Hmmm that's very interesting... What about `copy`? You know the bounds on the destination and source