@ksw2000 2024-06-19
The performance of `copy()` is still not better than `for-loop`.

```diff
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
+	copy(list.elements[l:], values)
-	for i := range values {
-		list.elements[l+i] = values[i]
-	}
}
```

```
goos: linux
goarch: amd64
pkg: github.com/emirpasic/gods/v2/lists/arraylist
cpu: Intel(R) Core(TM) i7-4790 CPU @ 3.60GHz
                     │ for-loop.txt │               copy.txt               │
                     │    sec/op    │    sec/op     vs base                │
ArrayListAdd100-8       788.6n ± 1%   1162.5n ± 1%  +47.41% (p=0.000 n=30)
ArrayListAdd1000-8      7.732µ ± 2%   10.980µ ± 2%  +42.02% (p=0.000 n=30)
ArrayListAdd10000-8     76.45µ ± 0%   114.61µ ± 1%  +49.92% (p=0.000 n=30)
ArrayListAdd100000-8    764.5µ ± 2%   1105.4µ ± 2%  +44.59% (p=0.000 n=30)
geomean                 24.43µ         35.66µ       +45.95%

                     │ for-loop.txt │               copy.txt               │
                     │     B/op     │     B/op      vs base                │
ArrayListAdd100-8      2.075Ki ± 1%   2.712Ki ± 1%  +30.68% (p=0.000 n=30)
ArrayListAdd1000-8     24.15Ki ± 3%   29.42Ki ± 3%  +21.83% (p=0.000 n=30)
ArrayListAdd10000-8    204.5Ki ± 1%   267.4Ki ± 1%  +30.79% (p=0.000 n=30)
ArrayListAdd100000-8   2.307Mi ± 4%   2.887Mi ± 1%  +25.17% (p=0.000 n=30)
geomean                70.14Ki        89.12Ki       +27.06%

                     │ for-loop.txt │              copy.txt               │
                     │  allocs/op   │ allocs/op   vs base                 │
ArrayListAdd100-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd1000-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd10000-8    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
ArrayListAdd100000-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=30) ¹
geomean                           ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```