@PapaCharlie 2024-06-19
There's probably something else at play here. Can you put your copy and append code here? And when you run the benchmarks, can you run the profiler (with `-cpuprofile`) and show the results here? You can get some graphs with `go tool pprof`