@PapaCharlie 2024-06-19
Oh and can you show your benchmarks? All of these will behave differently depending on how you're benchmarking them. For example, how many elements are you adding, etc