@ksw2000 2024-06-19
I use the benchmark in `/lists/arraylist/arraylist_test.go`

+ ArrayListAdd100 adds 100 elements into arrayList
+ ArrayListAdd1000 adds 1000 elements into arrayList
+ ArrayListAdd10000 adds 10000 elements into arrayList
+ ArrayListAdd100000 adds 100000 elements into arrayList

```go
func benchmarkAdd(b *testing.B, list *List[int], size int) {
	for i := 0; i < b.N; i++ {
		for n := 0; n < size; n++ {
			list.Add(n)
		}
	}
}

func BenchmarkArrayListAdd100(b *testing.B) {
	b.StopTimer()
	size := 100
	list := New[int]()
	b.StartTimer()
	benchmarkAdd(b, list, size)
}
```

The **append** method
```go
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
	list.elements = append(list.elements[:l], values...)
}
```

The **copy** method
```go
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
	copy(list.elements[l:], values)
}
```

I will run the benchmark again using a profiler (with `-cpuprofile`) and show the results here later.