@ksw2000 2024-06-19
The **for-loop** method
```go
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
	for i := range values {
		list.elements[l+i] = values[i]
	}
}
```
![for](https://github.com/emirpasic/gods/assets/13825170/0790fd02-bc1a-4095-8e86-a4f8a8bc3174)

The **append** method
```go
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
	list.elements = append(list.elements[:l], values...)
}
```

![append](https://github.com/emirpasic/gods/assets/13825170/65acf3f9-6072-479b-8731-853f2f889e04)


The **copy** method
```go
func (list *List[T]) Add(values ...T) {
	l := len(list.elements)
	list.growBy(len(values))
	copy(list.elements[l:], values)
}
```
![copy](https://github.com/emirpasic/gods/assets/13825170/944e9b01-f34b-400c-a8a5-627a39bbb130)

---

The left part is the **append** method, and the right part is the **for-loop** method. Maybe the `memmove` caused the difference between these two implementations. Similarly, implementing with **copy** also leads to a path that involves `memmove`.

<img width="305" alt="image" src="https://github.com/emirpasic/gods/assets/13825170/4903d847-1980-4b93-913f-d3e79d964d6d">

If `runtime.memmove` appears frequently in a CPU profile, it suggests that a lot of memory copying operations are occurring, which can be a performance bottleneck. This often happens in programs that do a lot of slice manipulations, such as frequent appends or deletes.